+++ src/edit/holdings/holdingsStatementFields.js
@@ -1,11 +1,12 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { TextArea } from '@folio/stripes/components';
 
 import RepeatableField from '../../components/RepeatableField';
 
-const HoldingsStatementFields = () => (
+const HoldingsStatementFields = ({ canAdd, canEdit }) => (
   <RepeatableField
     name="holdingsStatements"
     label={<FormattedMessage id="ui-inventory.holdingsStatements" />}
@@ -31,7 +32,19 @@ const HoldingsStatementFields = () => (
         rows: 1,
       },
     ]}
+    canAdd={canAdd}
+    canEdit={canEdit}
   />
 );
+
+HoldingsStatementFields.propTypes = {
+  canAdd: PropTypes.bool,
+  canEdit: PropTypes.bool,
+};
+
+HoldingsStatementFields.defaultProps = {
+  canAdd: true,
+  canEdit: true,
+};
 
 export default HoldingsStatementFields;