+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * Edit MARC Holdings via quickMARC. Refs UIIN-1636.
 * Instance record: Update Instance record Actions menu. Refs UIIN-1625.
 * Add ability to move MARC holdings between instances. Refs UIIN-1633.
+* Make mapped fields read-only for MARC holdings records. Refs UIIN-1639.
 
 ## [7.1.4](https://github.com/folio-org/ui-inventory/tree/v7.1.4) (2021-08-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.3...v7.1.4)